cleaned up util service, moved shouldresize into CacheService and made the check internal, moved json child serializer creation to json utils
