+++ src/CSharp App/Json/Converters/ColorToColorArrayConverter.cs
@@ -1,6 +1,7 @@
 using System;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
+using VolumetricSelection2077.Json.Helpers;
 using VolumetricSelection2077.Services;
 
 namespace VolumetricSelection2077.Json.Converters;
@@ -24,7 +25,7 @@ public class ColorToColorArrayConverter : JsonConverter<float[]>
                     throw new JsonSerializationException("Color object must contain r, g, and b properties.");
                 return new float[] { obj["r"].Value<float>() / 255f, obj["g"].Value<float>() / 255f, obj["b"].Value<float>() / 255f };
             case JsonToken.StartArray:
-                var cleanSerializer = UtilService.CreateChildSerializer(serializer, typeof(ColorToColorArrayConverter));
+                var cleanSerializer = JsonSerializerUtils.CreateChildSerializer(serializer, typeof(ColorToColorArrayConverter));
                 var arr = JArray.Load(reader);
                 return arr.ToObject<float[]>(cleanSerializer);
             default:
@@ -34,7 +35,7 @@ public class ColorToColorArrayConverter : JsonConverter<float[]>
 
     public override void WriteJson(JsonWriter writer, float[]? value, JsonSerializer serializer)
     {
-        var cleanSerializer = UtilService.CreateChildSerializer(serializer, typeof(ColorToColorArrayConverter));
+        var cleanSerializer = JsonSerializerUtils.CreateChildSerializer(serializer, typeof(ColorToColorArrayConverter));
         cleanSerializer.Serialize(writer, value);
     }
 }
