+++ src/CSharp App/Json/Helpers/JsonSerializerUtils.cs
@@ -1,41 +1,51 @@
 using System;
+using System.Linq;
 using Newtonsoft.Json;
 
-namespace VolumetricSelection2077.Helpers;
+namespace VolumetricSelection2077.Json.Helpers;
 
 public class JsonSerializerUtils
 {
-    public static JsonSerializer CloneWithoutConverters(JsonSerializer original)
+    /// <summary>
+    /// Creates a child serializer that excludes the given converters
+    /// </summary>
+    /// <param name="parent">serializer to copy settings from</param>
+    /// <param name="exclude">converters to exclude</param>
+    /// <returns></returns>
+    public static JsonSerializer CreateChildSerializer(JsonSerializer parent, params Type[] exclude)
     {
-        ArgumentNullException.ThrowIfNull(original);
-
-        return new JsonSerializer
+        var child = new JsonSerializer
         {
-            Formatting = original.Formatting,
-            Culture = original.Culture,
-            CheckAdditionalContent = original.CheckAdditionalContent,
-            ConstructorHandling = original.ConstructorHandling,
-            Context = original.Context,
-            ContractResolver = original.ContractResolver,
-            DateFormatHandling = original.DateFormatHandling,
-            DateFormatString = original.DateFormatString,
-            DateParseHandling = original.DateParseHandling,
-            DateTimeZoneHandling = original.DateTimeZoneHandling,
-            DefaultValueHandling = original.DefaultValueHandling,
-            EqualityComparer = original.EqualityComparer,
-            FloatFormatHandling = original.FloatFormatHandling,
-            FloatParseHandling = original.FloatParseHandling,
-            MaxDepth = original.MaxDepth,
-            MetadataPropertyHandling = original.MetadataPropertyHandling,
-            MissingMemberHandling = original.MissingMemberHandling,
-            NullValueHandling = original.NullValueHandling,
-            ObjectCreationHandling = original.ObjectCreationHandling,
-            PreserveReferencesHandling = original.PreserveReferencesHandling,
-            ReferenceLoopHandling = original.ReferenceLoopHandling,
-            StringEscapeHandling = original.StringEscapeHandling,
-            TraceWriter = original.TraceWriter,
-            TypeNameAssemblyFormatHandling = original.TypeNameAssemblyFormatHandling,
-            TypeNameHandling = original.TypeNameHandling
+            Culture = parent.Culture,
+            ContractResolver = parent.ContractResolver,
+            CheckAdditionalContent = parent.CheckAdditionalContent,
+            ConstructorHandling = parent.ConstructorHandling,
+            Context = parent.Context,
+            DateFormatHandling = parent.DateFormatHandling,
+            DateFormatString = parent.DateFormatString,
+            DateParseHandling = parent.DateParseHandling,
+            DateTimeZoneHandling = parent.DateTimeZoneHandling,
+            DefaultValueHandling = parent.DefaultValueHandling,
+            EqualityComparer = parent.EqualityComparer,
+            FloatFormatHandling = parent.FloatFormatHandling,
+            FloatParseHandling = parent.FloatParseHandling,
+            Formatting = parent.Formatting,
+            MaxDepth = parent.MaxDepth,
+            MetadataPropertyHandling = parent.MetadataPropertyHandling,
+            MissingMemberHandling = parent.MissingMemberHandling,
+            NullValueHandling = parent.NullValueHandling,
+            ObjectCreationHandling = parent.ObjectCreationHandling,
+            PreserveReferencesHandling = parent.PreserveReferencesHandling,
+            ReferenceLoopHandling = parent.ReferenceLoopHandling,
+            StringEscapeHandling = parent.StringEscapeHandling,
+            TypeNameAssemblyFormatHandling = parent.TypeNameAssemblyFormatHandling,
+            TypeNameHandling = parent.TypeNameHandling
         };
+
+        foreach (var c in parent.Converters)
+            if (!exclude.Contains(c.GetType()))
+                child.Converters.Add(c);
+
+        return child;
     }
 }
