+++ src/CSharp App/Services/CacheServiceRW.cs
@@ -258,8 +258,8 @@ public partial class CacheService
 
         if (!_settings.CacheEnabled)
         {
-            ClearDatabase(CacheDatabases.Vanilla, UtilService.ShouldResize(CacheDatabases.Vanilla, GetStats(), _settings.CacheDirectory));
-            ClearDatabase(CacheDatabases.Modded, UtilService.ShouldResize(CacheDatabases.Modded, GetStats(), _settings.CacheDirectory));
+            ClearDatabase(CacheDatabases.Vanilla, true);
+            ClearDatabase(CacheDatabases.Modded, true);
         }
